Server/requirements.txt
Flask
flask-cors
gunicorn
psycopg2-binary
Flask-SQLAlchemy
openai
python-dotenv
pgvector
supabase
pandas